Devel

+++ frontend/docker_entrypoint.sh
@@ -8,19 +8,33 @@ cd /home/cert/customer-portal
 
 npm install
 
-PATH=$(npm bin):$PATH bower install
+npx bower install
 
-echo 'create config/envs/devel.json.'
-cp config/envs/docker.json config/envs/devel.json
+if [ ! -f config/envs/devel.json ]; then
+  echo 'create docker devel config'
+  cp config/envs/docker.json config/envs/devel.json
+else
+  if ! cmp --silent config/envs/devel.json config/envs/docker.json
+  then
+    echo 'Warning: config/envs/devel.json differs from config/envs/docker.json!'
+  fi
+fi
 
-echo 'create config/envs/production.json.'
-cp config/envs/docker.json config/envs/production.json
+if [ ! -f config/envs/production.json ]; then
+  echo 'create docker production config'
+  cp config/envs/docker.json config/envs/production.json
+else
+  if ! cmp --silent config/envs/production.json config/envs/docker.json
+  then
+    echo 'Warning: config/envs/production.json differs from config/envs/docker.json!'
+  fi
+fi
 
 if [ ! -f secret.json ]; then
     echo 'create secret.json.'
     cp secret.json.example secret.json
 fi
 
-PATH=$(npm bin):$PATH grunt serve
+npx grunt serve
 
 } >&2 # redirect stout to stderr
